+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 ------------------
 
 - #1845 Added edit form adapter for lab contacts
+- #1846 Fix UnicodeDecodeError on Attachment upload
+- #1847 Added Analysis Profile Edit Form Adapter
 - #1844 Fix instrument results import
 - #1842 Refactor instrument results import form
 - #1841 Do not allow client user to deactivate its own client